Factor out functions for handling parsing/resizing of base64images to api utils
